requirements.txt
feedparser==6.0.11
markdown==3.5.2
PyYAML==6.0.1
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
beautifulsoup4==4.12.2
Jinja2==3.1.3
